requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.2.0
openai>=1.50.0
pydantic==2.5.3
pydantic-settings==2.1.0
supabase>=2.0.0
PyJWT>=2.8.0
